requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
motor==3.3.2
pymongo==4.6.1
python-multipart==0.0.9
python-dotenv==1.0.1
pydantic==2.7.4
PyPDF2==3.0.1
aiofiles==23.2.1
requests==2.31.0
Pillow==10.0.1
google-genai==1.32.0
pydantic-settings==2.6.1
